separate provider definition handling
Since the provider.go file was getting a bit long in the tooth, splits
out the handlers for provider definitions into its own file.
